requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
dateparser==1.1.4
defusedxml==0.7.1
Django==3.2.18
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-jet-reboot==1.3.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
feedparser==6.0.10
google-api-python-client==1.4.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.1
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
jet-bridge-base==1.3.1
jet-django==1.4.4
Markdown==3.4.1
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.3.0
promise==2.3
prompt-toolkit==2.0.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.1
sgmllib3k==1.0.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.5
zipp==3.13.0
